refactor(runtime-node): finish release issue decomposition
Keep the 0.8.1 release branch locally reviewable by finishing the runtime-node split and closing the exact verifier gaps for gateway token scope containment and memory backend plugins.

Constraint: Work stays local on release/v0.8.1 with no push or PR.

Rejected: Treat canMutateToken and MemoryBackendPlugin as sufficient by existence alone | the runtime routes and memory provider selection needed real integration.

Confidence: high

Scope-risk: moderate

Reversibility: clean

Directive: Keep index.ts as runtime orchestration; add new route or lifecycle logic to focused modules instead of growing the entrypoint.

Tested: npm run typecheck; npm test; npm run build:ui --workspace @crowclaw/web; npm run build:html --workspace @crowclaw/web

Related: #74 #90 #155
